re-org load tests
